test: isolate pricing page test target
